+++ source/Messaging.Application/Transactions/MoveIn/ForwardMeteringPointMasterDataHandler.cs
@@ -17,7 +17,6 @@ using System.Globalization;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
-using Messaging.Application.Common;
 using Messaging.Application.Configuration;
 using Messaging.Application.MasterData;
 using Messaging.Application.OutgoingMessages;
