+++ source/Messaging.Application/Transactions/MoveIn/MoveInNotifications.cs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 using System;
-using Messaging.Application.Common;
 using Messaging.Application.Configuration;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Application.OutgoingMessages.Common;
