+++ source/Messaging.ArchitectureTests/Application/InternalCommandTests.cs
@@ -17,7 +17,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Text.Json.Serialization;
-using Messaging.Application.Common.Commands;
+using Messaging.Application.Configuration.Commands.Commands;
 using Messaging.Infrastructure.Configuration;
 using Xunit;
