+++ source/Messaging.Infrastructure/Configuration/CompositionRoot.cs
@@ -25,9 +25,9 @@ using Energinet.DataHub.MessageHub.Client.Storage;
 using Energinet.DataHub.MessageHub.Model.Peek;
 using MediatR;
 using MediatR.Registration;
-using Messaging.Application.Common;
 using Messaging.Application.Configuration;
 using Messaging.Application.Configuration.Authentication;
+using Messaging.Application.Configuration.Commands;
 using Messaging.Application.Configuration.DataAccess;
 using Messaging.Application.IncomingMessages.RequestChangeOfSupplier;
 using Messaging.Application.MasterData.MarketEvaluationPoints;
