+++ source/Messaging.Infrastructure/Configuration/InternalCommands/CommandScheduler.cs
@@ -14,8 +14,9 @@
 
 using System;
 using System.Threading.Tasks;
-using Messaging.Application.Common.Commands;
 using Messaging.Application.Configuration;
+using Messaging.Application.Configuration.Commands;
+using Messaging.Application.Configuration.Commands.Commands;
 using Messaging.Infrastructure.Configuration.DataAccess;
 using Messaging.Infrastructure.Configuration.Serialization;
 using NodaTime;
