+++ source/Messaging.Infrastructure/Configuration/InternalCommands/CommandSchedulerFacade.cs
@@ -13,8 +13,9 @@
 // limitations under the License.
 
 using System.Threading.Tasks;
-using Messaging.Application.Common.Commands;
 using Messaging.Application.Configuration;
+using Messaging.Application.Configuration.Commands;
+using Messaging.Application.Configuration.Commands.Commands;
 using Messaging.Application.Configuration.DataAccess;
 
 namespace Messaging.Infrastructure.Configuration.InternalCommands;
