+++ source/Messaging.Infrastructure/Configuration/Processing/UnitOfWorkBehaviour.cs
@@ -16,8 +16,8 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
-using Messaging.Application.Common.Commands;
 using Messaging.Application.Configuration;
+using Messaging.Application.Configuration.Commands.Commands;
 using Messaging.Application.Configuration.DataAccess;
 using Messaging.Infrastructure.Configuration.DataAccess;
 using Microsoft.EntityFrameworkCore;
