+++ source/Messaging.Infrastructure/OutgoingMessages/CharacteristicsOfACustomerAtAnAp/CharacteristicsOfACustomerAtAnApDocumentWriter.cs
@@ -16,7 +16,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Xml;
-using Messaging.Application.Common;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Application.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Messaging.Application.OutgoingMessages.Common;
