+++ source/Messaging.Infrastructure/OutgoingMessages/Common/Xml/DocumentWriter.cs
@@ -18,7 +18,6 @@ using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml;
-using Messaging.Application.Common;
 using Messaging.Application.OutgoingMessages.Common;
 using Messaging.Application.OutgoingMessages.Common.Xml;
 using Messaging.Domain.OutgoingMessages;
