+++ source/Messaging.Infrastructure/OutgoingMessages/Common/Xml/HeaderWriter.cs
@@ -15,7 +15,6 @@
 using System;
 using System.Threading.Tasks;
 using System.Xml;
-using Messaging.Application.Common;
 using Messaging.Application.OutgoingMessages.Common.Xml;
 using Messaging.Domain.OutgoingMessages;
