+++ source/Messaging.Infrastructure/OutgoingMessages/ConfirmRequestChangeOfSupplier/ConfirmChangeOfSupplierJsonDocumentWriter.cs
@@ -16,7 +16,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
-using Messaging.Application.Common;
 using Messaging.Application.OutgoingMessages.Common;
 using Messaging.Application.OutgoingMessages.ConfirmRequestChangeOfSupplier;
 using Messaging.Domain.OutgoingMessages;
