+++ source/Messaging.Infrastructure/OutgoingMessages/GenericNotification/GenericNotificationDocumentWriter.cs
@@ -16,7 +16,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Xml;
-using Messaging.Application.Common;
 using Messaging.Application.OutgoingMessages.Common;
 using Messaging.Application.OutgoingMessages.Common.Xml;
 using Messaging.Application.OutgoingMessages.GenericNotification;
