+++ source/Messaging.Infrastructure/OutgoingMessages/Requesting/MessageRequestNotifications.cs
@@ -16,6 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Messaging.Application.Configuration;
+using Messaging.Application.Configuration.Commands;
 using Messaging.Application.OutgoingMessages.Requesting;
 
 namespace Messaging.Infrastructure.OutgoingMessages.Requesting
