+++ source/Messaging.IntegrationTests/Application/Transactions/MoveIn/CreateEndOfSupplyNotificationTests.cs
@@ -14,7 +14,6 @@
 
 using System;
 using System.Threading.Tasks;
-using Messaging.Application.Common;
 using Messaging.Application.Configuration;
 using Messaging.Application.Configuration.DataAccess;
 using Messaging.Application.OutgoingMessages;
