+++ source/Messaging.IntegrationTests/Application/Transactions/MoveIn/ForwardCustomerMasterDataTests.cs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 using System.Threading.Tasks;
-using Messaging.Application.Common;
 using Messaging.Application.Configuration;
 using Messaging.Application.Configuration.DataAccess;
 using Messaging.Application.MasterData;
