+++ source/Messaging.IntegrationTests/Application/Transactions/MoveIn/ForwardMeteringPointMasterDataTests.cs
@@ -16,7 +16,6 @@ using System;
 using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
-using Messaging.Application.Common;
 using Messaging.Application.Configuration.DataAccess;
 using Messaging.Application.MasterData;
 using Messaging.Application.OutgoingMessages.AccountingPointCharacteristics;
