+++ source/Messaging.Tests/Application/OutgoingMessages/CharacteristicsOfACustomerAtAnAP/CharacteristicsOfACustomerAtAnApDocumentWriterTests.cs
@@ -17,7 +17,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Xml.Schema;
-using Messaging.Application.Common;
 using Messaging.Application.Configuration;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Application.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
