+++ source/Messaging.Tests/Application/OutgoingMessages/GenericNotificationDocumentWriterTests.cs
@@ -19,7 +19,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 using System.Xml.Schema;
-using Messaging.Application.Common;
 using Messaging.Application.Configuration;
 using Messaging.Application.OutgoingMessages.Common;
 using Messaging.Application.OutgoingMessages.GenericNotification;
